+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @convergence-rfq/cli
 
+## 4.0.23-rc.5
+
+### Patch Changes
+
+- fix fixedSize, response decimals, reduce rpc calls
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.23
+
+## 4.0.23-rc.4
+
+### Patch Changes
+
+- add oracle id to psyoptions euro creation
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.23
+
+## 4.0.23-rc.3
+
+### Patch Changes
+
+- update cli script
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.23
+
+## 4.0.23-rc.2
+
+### Patch Changes
+
+- add psyoptions american program creation method, CvgWallet
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.25
+
+## 4.0.24
+
+### Patch Changes
+
+- add psyoptions american program creation method, CvgWallet
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.24
+
+## 4.0.24-rc
+
+### Patch Changes
+
+- update version number to have `rc`
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.24
+
+## 4.0.23
+
+### Patch Changes
+
+- fix collateral locked tokens amount
+- Updated dependencies
+  - @convergence-rfq/sdk@4.0.23
+
 ## 4.0.22
 
 ### Patch Changes